Scrapped month old code for old working version
I finally decided to give up on the old code and go back to when everything was working. The new script is titled "AIControllerFIXED" and it is currently a single script, not even a state machine. I *may* switch to a state machine in the future once I'm happy with the state (pun intended) of the code in its execution.

For now though, the AI at least moves mostly well; it does a fine job of staying on my hand(s). It just has a bit of jitter thanks to the way I'm setting the rotation of the cube to the normals of the surface it's standing on. https://gfycat.com/ExcellentDefiniteGosling